Move estimate_gas_for_builder_tx and signed_builder_tx to helpers
