+++ crates/op-rbuilder/src/primitives/mod.rs
@@ -1,7 +1,10 @@
 //! This module contains types from the reth that weren't heavily modified
 mod execution;
+mod helpers;
 mod payload_builder_service;
 
 pub use payload_builder_service::PayloadBuilderService;
+
+pub use helpers::{estimate_gas_for_builder_tx, signed_builder_tx};
 
 pub use execution::{ExecutedPayload, ExecutionInfo};
