+++ crates/op-rbuilder/src/tester/mod.rs
@@ -24,7 +24,6 @@ use reth_node_api::{EngineTypes, PayloadTypes};
 use reth_optimism_node::OpEngineTypes;
 use reth_payload_builder::PayloadId;
 use reth_rpc_layer::{AuthClientLayer, AuthClientService, JwtSecret};
-use serde_json;
 use serde_json::Value;
 use std::str::FromStr;
 use std::time::SystemTime;
